+++ src/games/planetwars/client/app/components/Home.scss
@@ -3,7 +3,7 @@
     display: flex;
     flex-direction: row;
     padding-left: 10px;
-    .configController {
+    .gameSetup {
       width: 40%;
       padding: 10px 10px 10px 10px;
     }
