+++ src/games/planetwars/client/app/components/gameSetup/configSelector/ConfigSelector.scss
