+++ src/games/planetwars/client/app/components/gameSetup/configform/ConfigForm.scss
